tuplex

Trait FromTuple

pub trait FromTuple<Tup> {
    // Required method
    fn from_tuple(tup: Tup) -> Self;
}

Converts a tuple from another one, the fields of which can be converted into the fields of the new tuple.

Required Methods

fn from_tuple(tup: Tup) -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.