tuplex

Trait HomoTuple

pub trait HomoTuple<T>
where
    Self: Len + IntoArray<T> + IntoBoxedSlice<T>,
{
    type FieldsWrapped;
    type TupleWrapped: IntoIterator<Item = T>;

    // Required methods
    fn wrap_fields(self) -> Self::FieldsWrapped;
    fn wrap_tuple(self) -> Self::TupleWrapped;
    fn wrap_into_iter(self) -> <Self::TupleWrapped as IntoIterator>::IntoIter;
}

Homogeneous Tuple’s trait.

Required Associated Types

type FieldsWrapped

New type of tuple after fields converted into Option.

type TupleWrapped: IntoIterator<Item = T>

New type of tuple after the whole tuple has been wrapped.

Required Methods

fn wrap_fields(self) -> Self::FieldsWrapped

Converts fields into Option.

fn wrap_tuple(self) -> Self::TupleWrapped

Wraps the whole tuple and get a new type.

fn wrap_into_iter(self) -> <Self::TupleWrapped as IntoIterator>::IntoIter

Converts the tuple into an iterater which owns the fields, by internally converting all fields into Options.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.