tuplex

Trait IntoArray

pub trait IntoArray<T> {
    type Output: Len;

    // Required method
    fn into_array(self) -> Self::Output;
}

Converts a tuple into an array, where the fields of the tuple can be converted into the same type of the array element.

Required Associated Types

type Output: Len

Required Methods

fn into_array(self) -> Self::Output