Trait tuplex::IntoBoxedSlice

pub trait IntoBoxedSlice<T> {
    // Required method
    fn into_boxed_slice(self) -> Box<[T]>;
}

Converts a tuple into a boxed slice, where the fields of the tuple can be converted into the same type of the slice element.

Required Methods

fn into_boxed_slice(self) -> Box<[T]>