Trait tuplex::IntoVec

pub trait IntoVec<T> {
    // Required method
    fn into_vec(self) -> Vec<T>;
}

Converts a tuple into a Vec, where the fields of the tuple can be converted into the same type of the Vec’s element.

Required Methods

fn into_vec(self) -> Vec<T>